minor bug fix

+++ preprocessing/to_freesurfer.py
@@ -97,7 +97,7 @@ def main(args):
         if args.surface:
             note('   - Projecting to surface...')
             (ldat, rdat) = sub.image_to_cortex(newimg, args.layer,
-                                               method=args.method, dtype=np.float)
+                                               method=args.method, dtype=np.float32)
             # we need to fix the dimensions...
             for (d,h) in zip([ldat,rdat], ['lh','rh']):
                 im = nib.freesurfer.mghformat.MGHImage(
